// webapp/types/product.d.ts
export interface IProduct {
    name: string,
    package: string, 
    price: number, 
    category: string,
    profile: string,
    pk: string,
    sk: string,
    userId: string,
    createdDate: string,
    pictures: string,
    description: string,
    featured: boolean,
    option: string,
    weight: string
  }